Black and function move
